Packet trace of IIT
The code is made to run for these "temp files. I am not considering the port numbers for the flow ids and just taking the IPs into account as of now. Also, there are some issues with the CAIDA dataset because of their large size. I might try to decrease their size using some software and then look up to them. So as of now, the "Heavy Elephant" problem as per the code written by them seems to be working for a dataset of my college